services/api/pyproject.toml
[project]
name = "pulse-api"
version = "0.1.0"
description = "Pulse FastAPI service"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.1",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.30.0",
  "fastapi>=0.115.7",
  "httpx>=0.28.1",
  "pydantic-settings>=2.7.1",
  "pyjwt[crypto]>=2.10.1",
  "sqlalchemy>=2.0.36",
  "uvicorn>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3",
  "ruff>=0.8.6"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

